+++ src/requirements/coins17.json
