+++ src/requirements/mast17.json
@@ -1,0 +1,105 @@
+{
+    "title": "卒業に必要な履修科目",
+    "description": "共通科目を除く",
+    "children": [
+        {
+            "title": "専門科目 - 必修科目",
+            "creditsCount": 12,
+            "courses": [
+                "GC41103", "GC41203", "GC48808", "GC48908"
+            ]
+        },
+        {
+            "title": "専門基礎科目 - 必修科目",
+            "creditsCount": 36,
+            "courses": [
+                "GC10112", "GC10122", "GC12101", "GC12103",
+                "GC11101", "GC11201", "GC11301", "GC11401",
+                "GC11501", "GC11601", "GC13101",
+                "GC12501", "GC12503", "GC12601", "GC12603",
+                "GC12301", "GC12303", "GC12401", "GC12403",
+                "GC15101", "GC15201", "GC13201",
+                "GC14301", "GC14401", "GC16101"
+            ]
+        },
+        {
+            "title": "専門科目 - 選択科目",
+            "description": "GC5",
+            "creditsCount": 30,
+            "courses": [
+                "GC50201", "GC50301", "GC50501", "GC50601",
+                "GC50701", "GC50801", "GC51001", "GC51101",
+                "GC51401", "GC51701", "GC51901", "GC52001",
+                "GC52101", "GC52201", "GC52301", "GC52401",
+                "GC52501", "GC52603", "GC52701", "GC52801",
+                "GC52901", "GC53303", "GC53401", "GC53502",
+                "GC53601", "GC53701", "GC53801", "GC53901",
+                "GC54001", "GC54101", "GC54201", "GC54301",
+                "GC54401", "GC59101", "GC59201", "GC59301",
+                "GC59401", "GC59501", "GC59601"
+            ]
+        },
+        {
+            "title": "専門基礎科目 - 選択科目",
+            "description": "GC2 / GA1",
+            "creditsCount": 22,
+            "courses": [
+                "GC20101", "GC20201", "GC21201", "GC21301",
+                "GC21401", "GC22101", "GC22201", "GC23101",
+                "GC23201", "GC23304", "GC23401", "GC23501",
+                "GC23601", "GC24101", "GC24312", "GC24401",
+                "GC25101", "GC26101", "GC27102",
+
+                "GA10101", "GA10201", "GA11102"
+            ]
+        },
+        {
+            "title": "自由科目",
+            "creditsCount": 9,
+            "children": [
+                {
+                    "title": "専門科目 - 自由科目",
+                    "description": "GC5 / GC6 / ビジネスシステムデザインA / ビジネスシステムデザインB",
+                    "creditsCount": 0,
+                    "courses": [
+                        "GC50201", "GC50301", "GC50501", "GC50601",
+                        "GC50701", "GC50801", "GC51001", "GC51101",
+                        "GC51401", "GC51701", "GC51901", "GC52001",
+                        "GC52101", "GC52201", "GC52301", "GC52401",
+                        "GC52501", "GC52603", "GC52701", "GC52801",
+                        "GC52901", "GC53303", "GC53401", "GC53502",
+                        "GC53601", "GC53701", "GC53801", "GC53901",
+                        "GC54001", "GC54101", "GC54201", "GC54301",
+                        "GC54401", "GC59101", "GC59201", "GC59301",
+                        "GC59401", "GC59501", "GC59601",
+
+                        "GA40303", "GA40403"
+                    ]
+                },
+                {
+                    "title": "専門基礎科目 - 自由科目",
+                    "description": "GC2 / GC3 / GA1",
+                    "creditsCount": 0,
+                    "courses": [
+                        "GC20101", "GC20201", "GC21201", "GC21301",
+                        "GC21401", "GC22101", "GC22201", "GC23101",
+                        "GC23201", "GC23304", "GC23401", "GC23501",
+                        "GC23601", "GC24101", "GC24312", "GC24401",
+                        "GC25101", "GC26101", "GC27102",
+
+                        "GC30401", "GC34502",
+
+                        "GA10101", "GA10201", "GA11102"
+                    ]
+                },
+                {
+                    "title": "基礎科目 - 自由科目",
+                    "description": "他学群又は多学類の授業科目 / 情報（上級） / 博物館に関する科目 / 自由科目（特設）",
+                    "creditsCount": 0,
+                    "courses": [],
+                    "allowOthers": true
+                }
+            ]
+        }
+    ]
+}
